requirements.txt
streamlit
pandas
duckdb
python-dotenv
requests
babel
duckdb-engine
plotly
